# Disable PostCSS-custom-properties in development