--- a/client/postcss.config.js
+++ b/client/postcss.config.js
@@ -1,19 +1,9 @@
-const config = require( './server/config' );
-const bundleEnv = config( 'env' );
-
-if ( bundleEnv === 'development' ) {
-	module.exports = () => ( {
-		plugins: {
-			autoprefixer: {},
-		},
-	} );
-} else {
-	module.exports = () => ( {
-		plugins: {
-			'postcss-custom-properties': {
-				importFrom: [ require.resolve( '@automattic/calypso-color-schemes' ) ],
-			},
-			autoprefixer: {},
-		},
-	} );
-}
+module.exports = ( { options } ) => {
+	const plugins = { autoprefixer: {} };
+	if ( options.transformCssProperties ) {
+		plugins[ 'postcss-custom-properties' ] = {
+			importFrom: [ require.resolve( '@automattic/calypso-color-schemes' ) ],
+		};
+	}
+	return { plugins };
+};