--- a/client/postcss.config.js
+++ b/client/postcss.config.js
@@ -1,8 +1,19 @@
-module.exports = () => ( {
-	plugins: {
-		'postcss-custom-properties': {
-			importFrom: [ require.resolve( '@automattic/calypso-color-schemes' ) ],
+const config = require( './server/config' );
+const bundleEnv = config( 'env' );
+
+if ( bundleEnv === 'development' ) {
+	module.exports = () => ( {
+		plugins: {
+			autoprefixer: {},
 		},
-		autoprefixer: {},
-	},
-} );
+	} );
+} else {
+	module.exports = () => ( {
+		plugins: {
+			'postcss-custom-properties': {
+				importFrom: [ require.resolve( '@automattic/calypso-color-schemes' ) ],
+			},
+			autoprefixer: {},
+		},
+	} );
+}