@alshakero 2020-06-01
Addressed all