@blowery 2020-06-01
Timings i saw for `yarn distclean; yarn; time yarn start`:

before: `yarn start  154.78s user 15.46s system 93% cpu 3:01.47 total`

after: `yarn start  158.08s user 15.82s system 190% cpu 1:31.14 total`

The "total" times here are likely a touch off, as I'm just control-c'ing my way out to get them. But in the right ballpark.

The main webpack build took `75346ms` on the branch, `132917ms` on the base commit. Very nice!

Looks like this feature came to postcss in `v8.0.10`. When did we pick that up?

Should also report this back up to the postcss folks.