@sgomes 2020-06-01
> There is a regression in IE. CSS vars are still there
> 
> https://hash-3b75df7a3b4cb9d293a6c3b934760d4f79fcf14d.calypso.live/calypso/fallback/entry-gutenboarding.cb2a1643ba37d01560e7.min.css

`browserslistEnv` will not have `fallback` but rather `default` for fallback builds, since that is the name of the `browserslist` configuration. That's likely the issue you're seeing.

The line should be changed to:

```js
transformCssProperties: browserslistEnv === 'defaults',
```