@sirbrillig 2020-06-01
Anecdotally, this cuts calypso build time for me from 6 minutes to 3.