@sgomes 2020-06-01
@nsakaimbo: Could you confirm that the desktop app still works correctly with these changes? We're looking for errors in styles, most likely the colours being wrong.