@nsakaimbo 2020-06-01
Desktop e2e failures in the build artifact - some sort of runtime error. It's happening consistently, although I don't think it's related to changes in this branch. Still looking...

`https://app.circleci.com/pipelines/github/Automattic/wp-desktop/14951/workflows/083203e2-0f53-4f50-87d9-731eff95c322/jobs/63125/artifacts`