@nsakaimbo 2020-06-01
@sgomes @alshakero This change looks great, except I'm unable to verify what it looks like as it seems latest Calypso isn't working in the desktop app (again - see the strange runtime error above). Unfortunately I'm unable to provide an assessment of whether there may be any unintended side effects here.