@jsnajdr 2020-06-02
I did some debugging and I think that the `postcss-custom-properties` plugin has some great optimization opportunities that could speed up even the production build significantly.

The real problem is not that the `isBlockIgnored` regexp check would be that inefficient, but rather that the surrounding code is called 700k+ times during the webpack build. The number of calls that are really needed is 3 🙂 

We use the `importFrom` option that loads some common CSS variables from a file in the `calypso-color-schemes` package. The `postcss-custom-properties` plugin reads and parses that file every time when creating an instance of the plugin.

And every CSS file needs its own instance of the plugin, because the PostCSS configuration is dynamic: every file can have a different chain of directory-local `postcss.config.js` files, and the config can be a function that returns a different config for each file.

Calypso webpack build creates 850 instances of the plugin, as that's the number of `*.scss` files to process. Every instance does its own `importFrom` processing, i.e., reading the file from disk, parsing it, ignoring the blocks that should be ignored, and extracting the found CSS variables into an object. The `getCustomPropertiesFromImports` results should be memoized, as they are always the same.

The `calypso-color-schemes.css` file has three `:root` selectors, hence the three necessary calls to `isBlockIgnored`.

Also, for every CSS variable in a block like this:
```css
:root {
  --studio-white: #fff;
  --studio-black: #000;
  /* ... 170 more variables like this ... */
}
```
the `isBlockIgnored` function is called on the _parent_ block (i.e., the `:root` selector), further inflating the number of calls up to hundreds of thousands.

A `postcss-custom-properties` patch that adds some memoizations and optimizations here and there could have a huge impact.