@alshakero 2020-06-02
Thanks for digging deep. This sounds great. If no one picked this up, I'll work on it on the weekend. I think we might need to create some sort of a benchmark to convince the plugin folks about the issue's significance to be able to compare before and after our PR.