@jsnajdr 2020-06-02
> I think we might need to create some sort of a benchmark to convince the plugin folks about the issue's significance.

My experience is that performance improvements usually don't need much convincing. Calypso build timings before and after should be enough.

There are two improvements that I expect to be impactful:

1. The more impactful one: in `getCustomPropertiesFromImports`, memoize the functions that convert a file name (string) to an object with custom properties, so that they actually read and process each file only once.
2. The less impactful one: try to reduce number of calls to `isBlockIgnored` by removing redundant calls where for each child, the parent node is tested repeatedly in a loop.