@jsnajdr 2020-06-02
@nsakaimbo I can't reproduce the runtime error in the `wp-desktop` app. Using the latest `develop` branch of `wp-desktop`, checking out the `update/disable-css-custom-props-in-dev` branch in the `calypso/` submodule, running `make` and opening the app in `releases/mac`. I can login and use the app just fine.

Do I need to go to any specific section to reproduce?