@ockham 2020-06-02
Thanks a ton @alshakero and @jsnajdr for your investigations! :sparkles: It'd be great if we could fix this upstream so that other A8c projects that use `postcss-custom-properties` through `calypso-build` (such as Jetpack) could also profit from the update.

(It would be otherwise conceivable to add an option to `calypso-build` to disable `postcss-custom-properties` for them, but that would require some hacks that are against the spirit of `calypso-build`, which looks for a build tool's own config file rather than adding its own wraparound option -- see [e.g.](https://github.com/Automattic/wp-calypso/blob/9000f926ccc67af288604471d2ea53c0befd54de/packages/calypso-build/README.md#advanced-usage-use-own-postcss-config))