@nsakaimbo 2020-06-02
Hi, @jsnajdr - I haven't been able to reproduce this locally, so at the moment I'm trying to "bisect" in CI: [link](https://github.com/Automattic/wp-desktop/compare/test-calypso-build)

I'm still testing, but right now the culprit seems to be Automattic/wp-calypso@d46489e.