@nsakaimbo 2020-06-02
Update: been trying to triage this and I think I might have narrowed down the break in the desktop app. Created a new issue with details: #42898

cc: @belcherj @griffbrad