@alshakero 2020-06-02
Thanks @nsakaimbo! Do I understand correctly that wp-desktop fails on Calypso#master too right? Can we merge this given it's inconsequential?