@sgomes 2020-06-03
> Thanks @nsakaimbo! Do I understand correctly that wp-desktop fails on Calypso#master too? Can we merge this given it's inconsequential to that issue?

Answering for @nsakaimbo: yes, `wp-desktop` currently fails on master, which means it shouldn't affect this issue at all. The `wp-desktop` failure is being tracked separately.