@alshakero 2020-06-03
Alright, so I created [an issue to triage the plugin](https://github.com/Automattic/wp-calypso/issues/42929) and I'll merge this.